--- akka-http-cors/src/main/java/ch/megard/akka/http/cors/japi/HttpHeaderRanges.java
@@ -1,8 +1,0 @@
-package ch.megard.akka.http.cors.japi;
-
-
-public final class HttpHeaderRanges {
-    private HttpHeaderRanges() { }
-
-    public static final HttpHeaderRange ALL = ch.megard.akka.http.cors.HttpHeaderRange.$times$.MODULE$;
-}
