+++ akka-http-cors/src/main/java/ch/megard/akka/http/cors/javadsl/model/HttpHeaderRanges.java
@@ -1,0 +1,8 @@
+package ch.megard.akka.http.cors.javadsl.model;
+
+
+public final class HttpHeaderRanges {
+    private HttpHeaderRanges() { }
+
+    public static final HttpHeaderRange ALL = ch.megard.akka.http.cors.scaladsl.model.HttpHeaderRange.$times$.MODULE$;
+}
